Debt: clean up localization contributions (#178861)
* refactor

* use async/await everywhere
